refactor: remove fake-timer on api call
