+++ package.json
@@ -15,6 +15,7 @@
     "coverage": "vitest run --coverage"
   },
   "dependencies": {
+    "@mswjs/interceptors": "^0.22.10",
     "@typescript-eslint/parser": "^5.54.1",
     "lodash": "^4.17.21",
     "react": "^18.2.0",
@@ -45,6 +46,7 @@
     "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-react": "^7.32.2",
     "jsdom": "^21.1.0",
+    "msw": "^1.2.1",
     "prettier": "2.8.4",
     "typescript": "^4.9.5",
     "vite": "^4.1.0",
