+++ src/tests/components/header.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { render, screen, within } from '@testing-library/react';
-import Header from '../components/header';
+import Header from '../../components/header';
 import { BrowserRouter } from 'react-router-dom';
 
 describe('Header', () => {
