+++ src/tests/pages/mainPage.test.tsx
@@ -1,8 +1,8 @@
 import { describe, it, expect } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import Header from '../components/header';
+import Header from '../../components/header';
 import { BrowserRouter } from 'react-router-dom';
-import SearchInput from '../components/searchInput';
+import SearchInput from '../../components/searchInput';
 
 describe('Main Page', () => {
   it('renders Header component', () => {
